Adding code coverage, more integration of functional tests with pytest access
