+++ .coveragerc
@@ -1,0 +1,5 @@
+[run]
+  branch = True
+  omit = */tests/*
+[report]
+  show_missing = True
