--- iocage/tests/0002_ioc_check_test.py
@@ -1,7 +1,0 @@
-from iocage.lib.ioc_check import IOCCheck
-
-
-def test_check():
-    IOCCheck()
-
-    assert True == True
