+++ iocage/tests/README.md
@@ -1,16 +1,62 @@
-# Running tests:
-================
-**These tests are written for `pytest`**
+# Code testing
+All the tests are written using the `pytest` unit testing framework. Code coverage is provided by `pytest-cov`
 
-- Make sure you're in the iocage directory
-- Supply --zpool="POOL" to pytest
-
-#####Example:
+Before running tests, test dependencies can be installed by running:
 ```
-git clone https://github.com/iocage/iocage.git
-cd iocage/iocage
-sudo pytest --zpool="TEST" --server="custom_server"
+$ pip install pytest-cov pytest-pep8
 ```
 
-**To list all supported fixtures to pytest:**
-`pytest --fixtures`
+## Unit tests
+
+Located in the ``tests/unit_tests`` directory, they can be started as a normal user with the following command:
+
+```
+$ pytest
+```
+
+## Functional tests
+
+Located in the ``tests/functional_tests``, they need a root acces and the name of a ZFS pool
+
+**/!\ The contents of the specified ZFS pool will be destroyed**
+
+To start the functional tests, run pytest with root privileges and the name of a zpool:
+```
+$ sudo pytest --zpool=mypool
+```
+
+Other parameters are available, to see them run:
+```
+$ pytest --fixtures
+```
+Extract:
+```
+zpool 
+    Specify a zpool to use.
+release 
+    Specify a RELEASE to use.
+server 
+    FTP server to login to.
+user 
+    The user to use for fetching.
+password 
+    The password to use for fetching.
+root_dir 
+    Root directory containing all the RELEASEs for fetching.
+http 
+    Have --server define a HTTP server instead.
+hardened 
+    Have fetch expect the default HardeneBSD layout instead.
+_file 
+    Use a local file directory for root-dir instead of FTP or HTTP.
+auth 
+    Authentication method for HTTP fetching. Valid values: basic, digest
+```
+
+
+#Example
+```
+$ git clone https://github.com/iocage/iocage.git
+$ cd iocage/iocage
+$ sudo pytest --zpool="TEST" --server="custom_server"
+```
