+++ iocage/tests/functional_tests/0002_ioc_check_test.py
@@ -1,0 +1,13 @@
+import pytest
+from iocage.lib.ioc_check import IOCCheck
+
+require_root = pytest.mark.require_root
+require_zpool = pytest.mark.require_zpool
+
+
+@require_root
+@require_zpool
+def test_check():
+    IOCCheck()
+
+    assert True == True
