+++ iocage/tests/functional_tests/0004_ioc_create_test.py
@@ -1,6 +1,12 @@
+import pytest
 from iocage.lib.ioc_create import IOCCreate
 
+require_root = pytest.mark.require_root
+require_zpool = pytest.mark.require_zpool
 
+
+@require_root
+@require_zpool
 def test_create(release, hardened):
     prop = ("tag=test",)
     prop_short = ("tag=test_short",)
