+++ iocage/tests/functional_tests/0005_ioc_start_test.py
@@ -1,9 +1,15 @@
+import pytest
 from iocage.lib.ioc_json import IOCJson
 from iocage.lib.ioc_list import IOCList
-from iocage.lib.ioc_stop import IOCStop
+from iocage.lib.ioc_start import IOCStart
+
+require_root = pytest.mark.require_root
+require_zpool = pytest.mark.require_zpool
 
 
-def test_stop():
+@require_root
+@require_zpool
+def test_start():
     jails, paths = IOCList("uuid").list_datasets()
 
     uuid = jails["test"]
@@ -15,7 +21,7 @@ def test_stop():
     conf = IOCJson(path).json_load()
     conf_short = IOCJson(path_short).json_load()
 
-    IOCStop(uuid, "test", path, conf)
-    IOCStop(uuid_short, "test_short", path_short, conf_short)
+    IOCStart(uuid, "test", path, conf)
+    IOCStart(uuid_short, "test_short", path_short, conf_short)
 
     assert True == True
