+++ iocage/tests/functional_tests/0008_ioc_get_test.py
@@ -1,7 +1,13 @@
+import pytest
 from iocage.lib.ioc_json import IOCJson
 from iocage.lib.ioc_list import IOCList
 
+require_root = pytest.mark.require_root
+require_zpool = pytest.mark.require_zpool
 
+
+@require_root
+@require_zpool
 def test_get():
     _, paths = IOCList("uuid").list_datasets()
