+++ iocage/tests/functional_tests/0009_ioc_destroy_test.py
@@ -1,7 +1,13 @@
+import pytest
 from iocage.lib.ioc_destroy import IOCDestroy
 from iocage.lib.ioc_list import IOCList
 
+require_root = pytest.mark.require_root
+require_zpool = pytest.mark.require_zpool
 
+
+@require_root
+@require_zpool
 def test_destroy():
     jails, paths = IOCList("uuid").list_datasets()
