+++ iocage/tests/functional_tests/9999_ioc_clean_test.py
@@ -1,8 +1,13 @@
 import os
-
+import pytest
 from iocage.lib.ioc_clean import IOCClean
 
+require_root = pytest.mark.require_root
+require_zpool = pytest.mark.require_zpool
 
+
+@require_root
+@require_zpool
 def test_clean():
     # Unless we change directory (not sure why) this will crash pytest.
     os.chdir("/")
