+++ setup.cfg
@@ -1,0 +1,4 @@
+[tool:pytest]
+addopts = -v -x -rs --ignore=setup.py --pep8 --cov-report term-missing --cov=iocage iocage iocage/tests
+pep8maxlinelength = 80
+pep8ignore = * ALL
